refactor(module:mention): refactor mention (#2510)
ref: #2381
