+++ components/mention/nz-mention.module.ts
@@ -3,9 +3,9 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { NzIconModule } from '../icon/nz-icon.module';
-import { NzMentionSuggestionDirective } from './mention-suggestions';
-import { NzMentionTriggerDirective } from './mention-trigger';
-import { NzMentionComponent } from './mention.component';
+import { NzMentionSuggestionDirective } from './nz-mention-suggestions';
+import { NzMentionTriggerDirective } from './nz-mention-trigger';
+import { NzMentionComponent } from './nz-mention.component';
 
 const COMPONENTS = [NzMentionComponent, NzMentionTriggerDirective, NzMentionSuggestionDirective];
