+++ components/mention/public-api.ts
@@ -1,3 +1,4 @@
-export * from './mention.module';
-export * from './mention.component';
-export * from './mention-trigger';
+export * from './nz-mention.module';
+export * from './nz-mention.component';
+export * from './nz-mention-trigger';
+export * from './nz-mention-suggestions';
